Fixed linting issues and consolidated fixtures files into the test files to be more standard golang
