--- pkg/gomason/publishing_fixtures.go
@@ -1,1 +1,0 @@
-package gomason
