+++ pkg/gomason/util_test.go
@@ -11,6 +11,51 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
+func testMetaDataJson() string {
+	return `{
+	"version": "0.1.0",
+	"package": "github.com/nikogura/testproject",
+	"description": "Test Project for Gomason.",
+	"building": {
+		"prepcommands": [
+      "echo \"GOPATH is: ${GOPATH}\""
+		],
+		"targets": [
+			{
+				"name": "linux/amd64"
+			},
+			{
+				"name": "darwin/amd64"
+			}
+		]
+	},
+	"signing": {
+		"program": "gpg",
+		"email": "gomason-tester@foo.com"
+	},
+	"publishing": {
+		"targets": [
+			{
+				"src": "testproject_darwin_amd64",
+				"dst": "{{.Repository}}/testproject/{{.Version}}/darwin/amd64/testproject",
+				"sig": true,
+				"checksums": true
+			},
+			{
+				"src": "testproject_linux_amd64",
+				"dst": "{{.Repository}}/testproject/{{.Version}}/linux/amd64/testproject",
+				"sig": true,
+				"checksums": true
+			}
+		]
+	}
+}`
+}
+
+func testMetadataFileName() string {
+	return "metadata.json"
+}
+
 func TestReadMetadata(t *testing.T) {
 	tmpDir, err := ioutil.TempDir("", "gomason")
 	if err != nil {
@@ -27,7 +72,7 @@ func TestReadMetadata(t *testing.T) {
 		t.Fail()
 	}
 
-	expected := TestMetadataObj()
+	expected := testMetadataObj()
 
 	actual, err := ReadMetadata(fileName)
 	if err != nil {
@@ -67,7 +112,7 @@ func TestParseStringForMetadata(t *testing.T) {
 }
 
 func TestGetCredentials(t *testing.T) {
-	username, password, err := GetCredentials(TestMetadataObj())
+	username, password, err := GetCredentials(testMetadataObj())
 	if err != nil {
 		log.Printf("Error getting credentials: %s", err)
 		t.Fail()
